packages/aid-py/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aid-discovery"
version = "1.1.0"
description = "Python library for Agent Interface Discovery (AID) - v1.1.0 release"
authors = [{ name = "Agent Community", email = "maintainers@agentcommunity.org" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = ["dnspython>=2.6.0", "idna>=3.6"]

[project.urls]
Homepage = "https://github.com/agent-community/agent-interface-discovery"
Repository = "https://github.com/agent-community/agent-interface-discovery"

[project.optional-dependencies]
dev = ["pytest>=8"]
# Preferred PKA backend (security-focused). Install to enable Ed25519 verification.
pka = ["cryptography>=42"]
